requirements.txt
asgiref==3.4.1
astroid==2.11.3
certifi==2021.10.8
charset-normalizer==2.0.10
defusedxml==0.7.1
dill==0.3.4
Django==3.2.11
django-extensions==3.1.5
django-filter==22.1
djangorestframework==3.13.1
httpie==2.6.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mysqlclient==2.1.0
pgi==0.0.11.2
pkg_resources==0.0.0
platformdirs==2.5.2
pydotplus==2.0.2
Pygments==2.11.2
pylint==2.13.7
pyparsing==3.0.8
PySocks==1.7.1
pytz==2021.3
requests==2.27.1
requests-toolbelt==0.9.1
sqlparse==0.4.2
tomli==2.0.1
typed-ast==1.5.3
typing_extensions==4.0.1
urllib3==1.26.7
wrapt==1.14.0
